folder refactor: helpers -> lib
