+++ src/lib/element.js
