+++ src/lib/emulation.js
