+++ src/lib/frame-load-recorder.js
