+++ src/lib/manifest-parser.js
