+++ src/lib/traces/devtools-timeline-model.js
