+++ src/lib/traces/processor.js
