+++ src/lib/web-inspector.js
