+++ test/helpers/browser/driver.js
@@ -17,8 +17,8 @@
 
 'use strict';
 
-const Driver = require('../../../helpers/drivers/cri.js');
-const Element = require('../../../helpers/element.js');
+const Driver = require('../../../lib/drivers/cri.js');
+const Element = require('../../../lib/element.js');
 const assert = require('assert');
 
 let DriverStub = new Driver();
