+++ test/helpers/element.js
@@ -17,7 +17,7 @@
 
 'use strict';
 
-const Element = require('../../helpers/element');
+const Element = require('../../lib/element');
 const assert = require('assert');
 
 class DriverStub {
